Moved application configuration to client configuration
Most configuration options need to vary by clients.
